backend/worker_celery/requirements.txt
absl-py==2.1.0
amqp==5.2.0
astunparse==1.6.3
async-timeout==4.0.3
billiard==4.2.0
cachetools==5.3.2
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.0
cycler==0.12.1
flatbuffers==23.5.26
fonttools==4.47.2
gast==0.5.4
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.1
h5py==3.10.0
idna==3.6
joblib==1.3.2
keras==2.15.0
kiwisolver==1.4.5
kombu==5.3.5
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
ml-dtypes==0.2.0
numpy==1.26.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.0
pillow==10.2.0
prompt-toolkit==3.0.43
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.35.0
termcolor==2.4.0
threadpoolctl==3.2.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.14.1
